% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCost.R
\name{estimasi_biaya_produksi}
\alias{estimasi_biaya_produksi}
\title{Estimasi Biaya Produksi
Menghitung total biaya produksi dan biaya per unit.}
\usage{
estimasi_biaya_produksi(bahan_baku, tenaga_kerja, overhead, jumlah_produksi)
}
\arguments{
\item{bahan_baku}{Numeric. Biaya bahan baku.}

\item{tenaga_kerja}{Numeric. Biaya tenaga kerja.}

\item{overhead}{Numeric. Biaya overhead.}

\item{jumlah_produksi}{Numeric. Total unit produksi.}
}
\value{
List dengan elemen:
  - `total_biaya_produksi`: Total biaya produksi.
  - `biaya_per_unit`: Biaya produksi per unit.
}
\description{
Estimasi Biaya Produksi
Menghitung total biaya produksi dan biaya per unit.
}
