% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCost.R
\name{hitung_bagi_hasil}
\alias{hitung_bagi_hasil}
\title{Hitung Bagi Hasil
Menghitung nilai bagi hasil berdasarkan persentase.}
\usage{
hitung_bagi_hasil(keuntungan_total, persentase)
}
\arguments{
\item{keuntungan_total}{Numeric. Total keuntungan.}

\item{persentase}{Numeric. Persentase bagi hasil.}
}
\value{
Numeric. Nilai bagi hasil.
}
\description{
Hitung Bagi Hasil
Menghitung nilai bagi hasil berdasarkan persentase.
}
