% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCost.R
\name{hitung_bep}
\alias{hitung_bep}
\title{Hitung Break-Even Point (BEP)
Menghitung jumlah unit yang harus terjual untuk mencapai BEP.}
\usage{
hitung_bep(biaya_tetap, biaya_variabel, harga_jual)
}
\arguments{
\item{biaya_tetap}{Numeric. Total biaya tetap.}

\item{biaya_variabel}{Numeric. Biaya variabel per unit.}

\item{harga_jual}{Numeric. Harga jual per unit.}
}
\value{
Numeric. Jumlah unit yang harus terjual untuk mencapai BEP.
}
\description{
Hitung Break-Even Point (BEP)
Menghitung jumlah unit yang harus terjual untuk mencapai BEP.
}
