% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCost.R
\name{hitung_cicilan}
\alias{hitung_cicilan}
\title{Hitung Cicilan Kredit
Menghitung cicilan bulanan untuk kredit.}
\usage{
hitung_cicilan(jumlah_pinjaman, bunga_tahunan, lama_pinjaman)
}
\arguments{
\item{jumlah_pinjaman}{Numeric. Total jumlah pinjaman.}

\item{bunga_tahunan}{Numeric. Suku bunga tahunan (dalam persen).}

\item{lama_pinjaman}{Integer. Lama pinjaman dalam bulan.}
}
\value{
Numeric. Besar cicilan bulanan.
}
\description{
Hitung Cicilan Kredit
Menghitung cicilan bulanan untuk kredit.
}
