% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCost.R
\name{hitung_diskon}
\alias{hitung_diskon}
\title{Hitung Diskon
Menghitung harga setelah diskon dan nilai diskon.}
\usage{
hitung_diskon(harga_awal, diskon_persen)
}
\arguments{
\item{harga_awal}{Numeric. Harga awal sebelum diskon.}

\item{diskon_persen}{Numeric. Persentase diskon.}
}
\value{
List dengan elemen:
  - `harga_diskon`: Harga setelah diskon.
  - `diskon`: Nilai diskon.
}
\description{
Hitung Diskon
Menghitung harga setelah diskon dan nilai diskon.
}
