% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCost.R
\name{hitung_pajak}
\alias{hitung_pajak}
\title{Kalkulator Pajak
Menghitung total pajak berdasarkan pendapatan dan tarif pajak.}
\usage{
hitung_pajak(pendapatan, tarif_pajak)
}
\arguments{
\item{pendapatan}{Numeric. Total pendapatan.}

\item{tarif_pajak}{Numeric. Tarif pajak (dalam persen).}
}
\value{
Numeric. Jumlah pajak yang harus dibayar.
}
\description{
Kalkulator Pajak
Menghitung total pajak berdasarkan pendapatan dan tarif pajak.
}
