% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCost.R
\name{hitung_total_belanja}
\alias{hitung_total_belanja}
\title{Hitung Total Belanja
Menghitung total belanja berdasarkan harga dan kuantitas barang.}
\usage{
hitung_total_belanja(harga, kuantitas)
}
\arguments{
\item{harga}{Numeric vector. Harga barang.}

\item{kuantitas}{Numeric vector. Kuantitas barang.}
}
\value{
Numeric. Total belanja.
}
\description{
Hitung Total Belanja
Menghitung total belanja berdasarkan harga dan kuantitas barang.
}
