% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCost.R
\name{hitung_untung_rugi}
\alias{hitung_untung_rugi}
\title{Hitung Untung Rugi
Menghitung total keuntungan atau kerugian dari penjualan barang.}
\usage{
hitung_untung_rugi(harga_beli, harga_jual, jumlah_barang)
}
\arguments{
\item{harga_beli}{Numeric. Harga beli per unit.}

\item{harga_jual}{Numeric. Harga jual per unit.}

\item{jumlah_barang}{Integer. Jumlah barang yang dijual.}
}
\value{
List dengan elemen:
  - `untung_rugi`: Total keuntungan atau kerugian.
  - `status`: Status "Untung" atau "Rugi".
}
\description{
Hitung Untung Rugi
Menghitung total keuntungan atau kerugian dari penjualan barang.
}
