% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCost.R
\name{simulasi_harga_jual}
\alias{simulasi_harga_jual}
\title{Simulasi Harga Jual
Menghitung harga jual dan margin keuntungan.}
\usage{
simulasi_harga_jual(biaya_per_unit, markup)
}
\arguments{
\item{biaya_per_unit}{Numeric. Biaya produksi per unit.}

\item{markup}{Numeric. Persentase markup untuk keuntungan.}
}
\value{
List dengan elemen:
  - `harga_jual`: Harga jual per unit.
  - `margin_keuntungan`: Margin keuntungan per unit.
}
\description{
Simulasi Harga Jual
Menghitung harga jual dan margin keuntungan.
}
